+++ crates/zng-view/src/image_cache.rs
@@ -1,21 +1,31 @@
-use std::{fmt, mem, sync::Arc};
+use std::{fmt, sync::Arc};
 
-use image::ImageDecoder;
+#[cfg(feature = "image_any")]
+use image::ImageDecoder as _;
+
 use webrender::api::{ImageDescriptor, ImageDescriptorFlags, ImageFormat};
 use winit::{
     event_loop::ActiveEventLoop,
     window::{CustomCursor, Icon},
 };
-use zng_txt::{ToTxt, Txt, formatx};
+#[cfg(feature = "image_any")]
+use zng_txt::ToTxt as _;
+use zng_txt::{Txt, formatx};
+
 use zng_unit::{Px, PxPoint, PxSize};
 use zng_view_api::{
     Event,
-    image::{ImageDataFormat, ImageDownscale, ImageId, ImageLoadedData, ImageMaskMode, ImagePpi, ImageRequest},
+    image::{ImageDataFormat, ImageId, ImageLoadedData, ImageMaskMode, ImagePpi, ImageRequest},
     ipc::{IpcBytes, IpcBytesReceiver},
 };
 
 use crate::{AppEvent, AppEventSender};
 use rustc_hash::FxHashMap;
+
+#[cfg(not(feature = "image_any"))]
+mod lcms2 {
+    pub struct Profile {}
+}
 
 pub(crate) const ENCODERS: &[&str] = &[
     #[cfg(feature = "image_png")]
@@ -127,6 +137,7 @@ impl ImageCache {
                             ),
                             mask,
                             ppi,
+                            None,
                         );
                         Ok((pixels, size, ppi, is_opaque, true))
                     } else {
@@ -148,6 +159,7 @@ impl ImageCache {
                             ),
                             None,
                             None,
+                            None,
                         );
                         Ok((pixels, size, None, is_opaque, false))
                     } else {
@@ -155,32 +167,41 @@ impl ImageCache {
                         Ok((data, size, None, is_opaque, true))
                     }
                 }
-                fmt => match Self::header_decode(&fmt, &data[..]) {
-                    Ok(h) => {
-                        let mut size = h.size;
-                        let decoded_len = size.width.0 as u64 * size.height.0 as u64 * 4;
-                        if decoded_len > max_decoded_len {
-                            Err(formatx!(
-                                "image {size:?} needs to allocate {decoded_len} bytes, but max allowed size is {max_decoded_len} bytes",
-                            ))
-                        } else {
-                            if let Some(d) = downscale {
-                                size = d.resize_dimensions(size);
-                            }
-                            let _ = app_sender.send(AppEvent::Notify(Event::ImageMetadataLoaded {
-                                image: id,
-                                size,
-                                ppi: h.ppi,
-                                is_mask: false,
-                            }));
-                            match Self::image_decode(&data[..], h.format, downscale, h.orientation) {
-                                Ok(img) => Ok(Self::convert_decoded(img, mask, h.ppi)),
-                                Err(e) => Err(e.to_txt()),
+                fmt => {
+                    #[cfg(not(feature = "image_any"))]
+                    {
+                        let _ = (max_decoded_len, downscale);
+                        Err(zng_txt::formatx!("no decoder for {fmt:?}"))
+                    }
+
+                    #[cfg(feature = "image_any")]
+                    match Self::header_decode(&fmt, &data[..]) {
+                        Ok(h) => {
+                            let mut size = h.size;
+                            let decoded_len = size.width.0 as u64 * size.height.0 as u64 * 4;
+                            if decoded_len > max_decoded_len {
+                                Err(formatx!(
+                                    "image {size:?} needs to allocate {decoded_len} bytes, but max allowed size is {max_decoded_len} bytes",
+                                ))
+                            } else {
+                                if let Some(d) = downscale {
+                                    size = d.resize_dimensions(size);
+                                }
+                                let _ = app_sender.send(AppEvent::Notify(Event::ImageMetadataLoaded {
+                                    image: id,
+                                    size,
+                                    ppi: h.ppi,
+                                    is_mask: false,
+                                }));
+                                match Self::image_decode(&data[..], h.format, downscale, h.orientation) {
+                                    Ok(img) => Ok(Self::convert_decoded(img, mask, h.ppi, h.icc_profile)),
+                                    Err(e) => Err(e.to_txt()),
+                                }
                             }
                         }
+                        Err(e) => Err(e),
                     }
-                    Err(e) => Err(e),
-                },
+                }
             };
 
             match r {
@@ -216,10 +237,11 @@ impl ImageCache {
             let mut full = vec![];
             let mut size = None;
             let mut ppi = None;
+            let mut icc_profile = None::<lcms2::Profile>;
             let mut is_encoded = true;
             let mut orientation = image::metadata::Orientation::NoTransforms;
 
-            let mut format = match &request_fmt {
+            let mut format: Option<image::ImageFormat> = match &request_fmt {
                 ImageDataFormat::Bgra8 { size: s, ppi: p } => {
                     is_encoded = false;
                     size = Some(*s);
@@ -231,7 +253,9 @@ impl ImageCache {
                     size = Some(*s);
                     None
                 }
+                #[cfg(feature = "image_any")]
                 ImageDataFormat::FileExtension(ext) => image::ImageFormat::from_extension(ext.as_str()),
+                #[cfg(feature = "image_any")]
                 ImageDataFormat::MimeType(t) => t.strip_prefix("image/").and_then(image::ImageFormat::from_extension),
                 ImageDataFormat::Unknown => None,
                 _ => None,
@@ -245,6 +269,7 @@ impl ImageCache {
 
                         full.extend(d);
 
+                        #[cfg(feature = "image_any")]
                         if let Some(fmt) = format {
                             if size.is_none() {
                                 if let Ok(h) = Self::header_decode(&request_fmt, &full) {
@@ -252,6 +277,7 @@ impl ImageCache {
                                     orientation = h.orientation;
                                     format = Some(h.format);
                                     ppi = h.ppi;
+                                    icc_profile = h.icc_profile;
                                 }
                                 if let Ok(mut d) = image::ImageReader::with_format(std::io::Cursor::new(&full), fmt).into_decoder() {
                                     use image::metadata::Orientation::*;
@@ -260,7 +286,7 @@ impl ImageCache {
                                     orientation = d.orientation().unwrap_or(NoTransforms);
 
                                     if matches!(orientation, Rotate90 | Rotate270 | Rotate90FlipH | Rotate270FlipH) {
-                                        mem::swap(&mut w, &mut h)
+                                        std::mem::swap(&mut w, &mut h)
                                     }
 
                                     size = Some(PxSize::new(Px(w as i32), Px(h as i32)));
@@ -289,9 +315,21 @@ impl ImageCache {
             }
 
             if let Some(fmt) = format {
+                #[cfg(not(feature = "image_any"))]
+                let _ = (
+                    fmt,
+                    max_decoded_len,
+                    downscale,
+                    mask,
+                    &mut icc_profile,
+                    &mut orientation,
+                    &mut format,
+                );
+
+                #[cfg(feature = "image_any")]
                 match Self::image_decode(&full[..], fmt, downscale, orientation) {
                     Ok(img) => {
-                        let (pixels, size, ppi, is_opaque, is_mask) = Self::convert_decoded(img, mask, ppi);
+                        let (pixels, size, ppi, is_opaque, is_mask) = Self::convert_decoded(img, mask, ppi, icc_profile);
                         let _ = app_sender.send(AppEvent::ImageLoaded(ImageLoadedData::new(
                             id, size, ppi, is_opaque, is_mask, pixels,
                         )));
@@ -357,6 +395,7 @@ impl ImageCache {
         let _ = self.app_sender.send(AppEvent::Notify(Event::ImageLoaded(data)));
     }
 
+    #[cfg(feature = "image_any")]
     fn header_decode(fmt: &ImageDataFormat, data: &[u8]) -> Result<ImageHeader, Txt> {
         let maybe_fmt = match fmt {
             ImageDataFormat::FileExtension(ext) => image::ImageFormat::from_extension(ext.as_str()),
@@ -394,10 +433,14 @@ impl ImageCache {
                 let (mut w, mut h) = decoder.dimensions();
                 let orientation = decoder.orientation().unwrap_or(NoTransforms);
                 if matches!(orientation, Rotate90 | Rotate270 | Rotate90FlipH | Rotate270FlipH) {
-                    mem::swap(&mut w, &mut h)
+                    std::mem::swap(&mut w, &mut h)
                 }
 
                 let mut ppi = None;
+                #[cfg(feature = "image_png")]
+                let mut png_gamma = None;
+                #[cfg(feature = "image_png")]
+                let mut png_chromaticities = None;
 
                 match fmt {
                     #[cfg(feature = "image_jpeg")]
@@ -451,7 +494,8 @@ impl ImageCache {
                     image::ImageFormat::Png => {
                         let d = png::Decoder::new_with_limits(std::io::Cursor::new(data), png::Limits { bytes: usize::MAX });
                         let d = d.read_info().map_err(|e| e.to_txt())?;
-                        if let Some(d) = d.info().pixel_dims {
+                        let info = d.info();
+                        if let Some(d) = info.pixel_dims {
                             match d.unit {
                                 png::Unit::Unspecified => {}
                                 png::Unit::Meter => {
@@ -459,6 +503,8 @@ impl ImageCache {
                                 }
                             }
                         }
+                        png_gamma = info.gama_chunk;
+                        png_chromaticities = info.chrm_chunk;
                     }
                     #[cfg(feature = "image_tiff")]
                     image::ImageFormat::Tiff => {
@@ -511,21 +557,36 @@ impl ImageCache {
                     }
                 }
 
+                let mut icc_profile = None;
+                if let Ok(Some(icc)) = decoder.icc_profile() {
+                    match lcms2::Profile::new_icc(&icc) {
+                        Ok(p) => icc_profile = Some(p),
+                        Err(e) => tracing::error!("error parsing ICC profile, {e}"),
+                    }
+                }
+                #[cfg(feature = "image_png")]
+                if icc_profile.is_none() {
+                    // PNG has some color management metadata, convert to standard
+                    icc_profile = crate::util::png_color_metadata_to_icc(png_gamma, png_chromaticities);
+                }
+
                 Ok(ImageHeader {
                     format: fmt,
                     size: PxSize::new(Px(w as i32), Px(h as i32)),
                     orientation,
                     ppi,
+                    icc_profile,
                 })
             }
             None => Err(Txt::from_static("unknown format")),
         }
     }
 
+    #[cfg(feature = "image_any")]
     fn image_decode(
         buf: &[u8],
         format: image::ImageFormat,
-        downscale: Option<ImageDownscale>,
+        downscale: Option<zng_view_api::image::ImageDownscale>,
         orientation: image::metadata::Orientation,
     ) -> image::ImageResult<image::DynamicImage> {
         let buf = std::io::Cursor::new(buf);
@@ -539,15 +600,16 @@ impl ImageCache {
 
         if let Some(s) = downscale {
             let (img_w, img_h) = (image.width(), image.height());
+            use zng_view_api::image::ImageDownscale::*;
             match s {
-                ImageDownscale::Fit(s) => {
+                Fit(s) => {
                     let w = img_w.min(s.width.0 as u32);
                     let h = img_h.min(s.height.0 as u32);
                     if w != img_w || h != img_h {
                         image = image.resize(w, h, image::imageops::FilterType::Triangle);
                     }
                 }
-                ImageDownscale::Fill(s) => {
+                Fill(s) => {
                     let w = img_w.min(s.width.0 as u32);
                     let h = img_h.min(s.height.0 as u32);
                     if w != img_w && h != img_h {
@@ -560,12 +622,17 @@ impl ImageCache {
         Ok(image)
     }
 
-    fn convert_decoded(image: image::DynamicImage, mask: Option<ImageMaskMode>, ppi: Option<ImagePpi>) -> RawLoadedImg {
+    fn convert_decoded(
+        image: image::DynamicImage,
+        mask: Option<ImageMaskMode>,
+        ppi: Option<ImagePpi>,
+        icc_profile: Option<lcms2::Profile>,
+    ) -> RawLoadedImg {
         use image::DynamicImage::*;
 
         let mut is_opaque = true;
 
-        let (size, pixels) = match image {
+        let (size, mut pixels) = match image {
             ImageLuma8(img) => (
                 img.dimensions(),
                 if mask.is_some() {
@@ -984,6 +1051,16 @@ impl ImageCache {
             _ => unreachable!(),
         };
 
+        #[cfg(feature = "image_any")]
+        if let Some(p) = icc_profile {
+            use lcms2::*;
+            let srgb = Profile::new_srgb();
+            let t = Transform::new(&p, PixelFormat::BGRA_8, &srgb, PixelFormat::BGRA_8, Intent::Perceptual).unwrap();
+            t.transform_in_place(&mut pixels);
+        }
+        #[cfg(not(feature = "image_any"))]
+        let _ = (icc_profile, &mut pixels);
+
         (
             IpcBytes::from_vec(pixels),
             PxSize::new(Px(size.0 as i32), Px(size.1 as i32)),
@@ -1040,12 +1117,13 @@ impl ImageCache {
         self.images.clear();
     }
 }
-
+#[cfg(feature = "image_any")]
 struct ImageHeader {
     format: image::ImageFormat,
     size: PxSize,
     orientation: image::metadata::Orientation,
     ppi: Option<ImagePpi>,
+    icc_profile: Option<lcms2::Profile>,
 }
 /// (pixels, size, ppi, is_opaque, is_mask)
 type RawLoadedImg = (IpcBytes, PxSize, Option<ImagePpi>, bool, bool);
@@ -1570,6 +1648,7 @@ mod capture {
                     ),
                     Some(mask),
                     ppi,
+                    None,
                 );
 
                 let id = self.add(ImageRequest::new(